+++ doc/conf.py
@@ -37,6 +37,7 @@ sys.path.insert(0, os.path.abspath('..'))
 # ones.
 extensions = [
     'celery.contrib.sphinx',
+    'matplotlib.sphinxext.plot_directive',
     'sphinx.ext.autodoc',
     'sphinx.ext.autosectionlabel',
     'sphinx.ext.graphviz',